# predictors/climate_risk_pred/swagger/swagger.yml
openapi: 3.0.3
info:
  title: Frost Predictor API
  description: API de predicción de riesgos agrícolas (heladas y plagas)
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Servidor de desarrollo
  - url: http://51.44.230.198:10000
    description: Servidor de producción
...
paths:
  /heladas/observadas/{tipo}:
    post:
      tags:
        - Heladas
      summary: Predicción de heladas observadas
      description: >
        Obtiene la predicción de riesgo de helada basada en datos históricos observados
        de los últimos 6 meses. Requiere indicar province O estacion, nunca ambos.
      parameters:
        - name: tipo
          in: path
          required: true
          description: Granularidad temporal de los datos históricos solicitados
          schema:
            type: string
            enum: [hora, dia, semana]
        - name: format
          in: query
          required: false
          description: Formato de respuesta. Si se indica pdf se devuelve informe firmado digitalmente
          schema:
            type: string
            enum: [json, pdf]
            default: json
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                codigo_provincia:
                  type: string
                  description: Código de provincia (excluyente con estacion)
                  example: "CC"
                codigo_estacion:
                  type: string
                  description: Código de estación meteorológica (excluyente con province)
                  example: "4452X"
                evaluacion:
                  type: string
                  description: Incluir evaluación de riesgo por variedades de cultivo
                  enum: ["true", "false", "1", "0", "yes", "si"]
                  default: "false"
                variedades:
                  type: string
                  description: Nombres de variedades separados por coma (requerido si evaluacion=true)
                  example: "merlot, tempranillo"
            examples:
              sinEvaluacion:
                summary: Petición básica por provincia
                value:
                  codigo_provincia: "CC"
                  evaluacion: "false"
              conEvaluacionVariedades:
                summary: Petición con evaluación de variedades
                value:
                  codigo_provincia: "CC"
                  evaluacion: "true"
                  variedades: "merlot, tempranillo"
      responses:
        "200":
          description: Predicción de heladas observadas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiesgoHeladaObservadaDTO'
            application/pdf:
              schema:
                type: string
                format: binary
                description: Informe PDF firmado digitalmente
        "400":
          $ref: '#/components/responses/BadRequest'
        "500":
          $ref: '#/components/responses/InternalError'
  /heladas/futuras/{zona}:
    post:
      tags:
        - Heladas
      summary: Predicción de heladas futuras por zona
      description: >
        Obtiene la predicción de riesgo de heladas futuras (mañana) para una zona geográfica.
        Permite filtrar por provincia O comunidad autónoma. Opcionalmente evalúa variedades
        de cultivo y/o localidades específicas.
      parameters:
        - name: zona
          in: path
          required: true
          description: Ámbito geográfico de la predicción
          schema:
            type: string
            enum: [provincial, ccaa, nacional]
          example: provincial
        - name: format
          in: query
          required: false
          description: Formato de respuesta. Si se indica pdf se devuelve informe firmado digitalmente
          schema:
            type: string
            enum: [json, pdf]
            default: json
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                codigo_provincia:
                  type: string
                  description: ID de la provincia (excluyente con ccaaId)
                  example: "CC"
                codigo_ccaa:
                  type: string
                  description: ID de la comunidad autónoma (excluyente con provinciaId)
                  example: "11"
                evaluacion_var:
                  type: boolean
                  description: Incluir evaluación de riesgo sobre variedades de cultivo
                  default: false
                evaluacion_loc:
                  type: boolean
                  description: Incluir evaluación de riesgo por localidades
                  default: false
                variedades:
                  type: array
                  description: Lista de variedades a evaluar (requerido si evaluacion_var=true)
                  items:
                    type: string
                  example: ["merlot", "tempranillo"]
                localidades:
                  type: array
                  description: Lista de localidades a evaluar (requerido si evaluacion_loc=true)
                  items:
                    type: string
                  example: ["plasencia", "cáceres"]
            examples:
              soloZona:
                summary: Predicción básica por provincia
                value:
                  codigo_provincia: "CC"
              conVariedades:
                summary: Con evaluación de variedades
                value:
                  codigo_provincia: "CC"
                  evaluacion_var: true
                  variedades: ["merlot", "tempranillo"]
              conLocalidades:
                summary: Con evaluación de localidades
                value:
                  codigo_provincia: "CC"
                  evaluacion_loc: true
                  localidades: ["plasencia", "cáceres"]
      responses:
        "200":
          description: Predicción de heladas futuras
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiesgoHeladaFuturaDTO'
            application/pdf:
              schema:
                type: string
                format: binary
                description: Informe PDF firmado digitalmente
        "400":
          $ref: '#/components/responses/BadRequest'
        "500":
          $ref: '#/components/responses/InternalError'
  /plagas/calculadas:
    post:
      tags:
        - Plagas
      summary: Predicción de riesgo de plagas calculadas
      description: >
        Obtiene el riesgo de plagas para la semana actual basándose en un calendario
        de alertas precalculado por ITACYL. No requiere datos climáticos en tiempo real.
        Devuelve una lista de RiesgoPlagaCultivoDTO con todas las plagas asociadas al cultivo.
      parameters:
        - name: cultivo
          in: query
          required: true
          description: Nombre del cultivo a analizar
          schema:
            type: string
          example: "vid"
        - name: format
          in: query
          required: false
          description: Formato de respuesta. Si se indica pdf se devuelve informe firmado digitalmente
          schema:
            type: string
            enum: [json, pdf]
            default: json
      responses:
        "200":
          description: Lista de riesgos de plagas calculados para el cultivo
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RiesgoPlagaCultivoDTO'
            application/pdf:
              schema:
                type: string
                format: binary
                description: Informe PDF firmado digitalmente
        "400":
          description: Parámetro cultivo no indicado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: "Invalid parameters"
                status: 400
                error: "Se debe indicar el valor de un query param : cultivo"
        "404":
          description: No se encontraron datos para el cultivo indicado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: "Data Not Found"
                status: 404
                error: "No se han encontrado datos para hacer la predicción de riesgos de plagas frente al cultivo : vid"
        "500":
          $ref: '#/components/responses/InternalError'
  /plagas/estimadas:
    post:
      tags:
        - Plagas
      summary: Predicción de riesgo de plagas estimadas
      description: >
        Obtiene la predicción de riesgo de plagas estimada para un cultivo en un rango
        de fechas, combinando datos de sensores IoT (opcionales) con datos meteorológicos del SiAR.
        Si se indican sensores, sus mediciones almacenadas tienen prioridad sobre los datos del SiAR.
        Devuelve la evaluación día a día de cada plaga seleccionada en el período indicado.
      parameters:
        - name: format
          in: query
          required: false
          description: Formato de respuesta. Si se indica pdf se devuelve informe firmado digitalmente
          schema:
            type: string
            enum: [json, pdf]
            default: json
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - cultivo
                - fecha_inicio
                - fecha_fin
              properties:
                cultivo:
                  type: string
                  description: Nombre del cultivo a analizar
                  example: "vid"
                fecha_inicio:
                  type: string
                  format: date-time
                  description: Inicio del período de evaluación (ISO 8601)
                  example: "2025-04-01T00:00:00Z"
                fecha_fin:
                  type: string
                  format: date-time
                  description: Fin del período de evaluación (ISO 8601)
                  example: "2025-04-30T00:00:00Z"
                id_plaga:
                  type: string
                  description: Identificador público de la plaga a evaluar
                  example: "PLAGA-TOMATE-01"
                datos_sensores:
                  type: array
                  description: >
                    Lista de sensores IoT cuyas mediciones almacenadas se quieren consultar.
                    Si se omite se usarán exclusivamente datos del SiAR.
                  items:
                    type: object
                    properties:
                      sensor: 
                        type: string
                        example: "a840419a0188cc08"
                      nombre_predictor_plaga: 
                        type: string
                        example: "humedad_foliar"
                parcela:
                  type: string
                  description: ID público de la parcela para contextualizar el cálculo (opcional)
                  example: "PARC-001"
                codigo_estacion:
                  type: string
                  description: Código identificador de la estación sobre la que se quiere recopilar datos climáticos históricos 
                    como salvaguarda para el análisis (opcional)
                  example: "CC01"
                codigo_provincia:
                  type: string
                  description: Código identificador de la provincia sobre la que se quiere recopilar datos climáticos históricos 
                    como salvaguarda para el análisis (opcional)
                  example: "CC"
            examples:
              conSensoresParcelaProvincia:
                summary: Con sensores IoT, parcela y provincias
                value:
                  cultivo: "vid"
                  fecha_inicio: "2025-04-01T00:00:00Z"
                  fecha_fin: "2025-04-30T00:00:00Z"
                  id_plaga: "PLAGA-TOMATE-01"
                  datos_sensores: [{"sensor" : "a840419a0188cc08", "nombre_predictor_plaga": "humedad_foliar"}]
                  parcela: "PARC-001"
                  codigo_provincia: "CC"
              conSensoresParcelaEstacion:
                summary: Con sensores IoT, parcela y estaciones meteorológicas
                value:
                  cultivo: "vid"
                  fecha_inicio: "2025-04-01T00:00:00Z"
                  fecha_fin: "2025-04-30T00:00:00Z"
                  id_plaga: "PLAGA-TOMATE-01"
                  datos_sensores: [{"sensor" : "a840419a0188cc08", "nombre_predictor_plaga": "humedad_foliar"}]
                  parcela: "PARC-001"
                  codigo_estacion: "CC01"
              conSensoresProvincia:
                summary: Con sensores IoT y provincias
                value:
                  cultivo: "vid"
                  fecha_inicio: "2025-04-01T00:00:00Z"
                  fecha_fin: "2025-04-30T00:00:00Z"
                  id_plaga: "PLAGA-TOMATE-01"
                  datos_sensores: [{"sensor" : "a840419a0188cc08", "nombre_predictor_plaga": "humedad_foliar"}]
                  codigo_provincia: "CC"
              conSensoresEstacion:
                summary: Con sensores IoT y estaciones meteorológicas
                value:
                  cultivo: "vid"
                  fecha_inicio: "2025-04-01T00:00:00Z"
                  fecha_fin: "2025-04-30T00:00:00Z"
                  id_plaga: "PLAGA-TOMATE-01"
                  datos_sensores: [{"sensor" : "a840419a0188cc08", "nombre_predictor_plaga": "humedad_foliar"}]
                  codigo_estacion: "CC01"
      responses:
        "200":
          description: Predicción estimada de plagas día a día
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiesgoPlagaEstimadaDTO'
            application/pdf:
              schema:
                type: string
                format: binary
                description: Informe PDF firmado digitalmente
        "400":
          $ref: '#/components/responses/BadRequest'
        "500":
          $ref: '#/components/responses/InternalError'
components:
  responses:
    BadRequest:
      description: Parámetros de la petición inválidos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: "Error interno del servidor"
            status: 500
            error: "Internal server error"
  schemas:
...
    Error:
      type: object
      properties:
        message:
          type: string
        status:
          type: integer
        error:
          type: string
    NivelRiesgo:
      type: string
      enum: [sin_riesgo, debil, moderada, fuerte]
...
    TipoPrecision:
      type: string
      enum: [alta, media, baja]
...
    RiesgoHeladaObservadaDTO:
      type: object
      description: Resultado de predicción de heladas basada en datos históricos observados
      properties:
        nivel:
          $ref: '#/components/schemas/NivelRiesgo'
        comentarios:
          type: string
          example: "Riesgo moderado detectado en base a registros de los últimos 6 meses"
        alertas:
          type: array
          items:
            $ref: '#/components/schemas/AlertaDTO'
        contexto:
          $ref: '#/components/schemas/ContextoCalculoDTO'
        tipo_prediccion:
          type: string
          enum: [hoy, manana, pasadomanana]
        evaluaciones_variedades:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/ResumenCultivoDTO'
        riesgos_heladas_blancas:
          type: array
          items:
            $ref: '#/components/schemas/RiesgoHeladaTipoDTO'
        riesgos_heladas_negras:
          type: array
          items:
            $ref: '#/components/schemas/RiesgoHeladaTipoDTO'
        fecha_comiezo_registros:
          type: string
          format: date
          example: "2024-10-26"
        fecha_fin_registros:
          type: string
          format: date
          example: "2025-04-26"
        registro_temperatura_minima:
          $ref: '#/components/schemas/RegistroTempMinDTO'
    RiesgoHeladaFuturaDTO:
      type: object
      description: Resultado de predicción de heladas basada en datos futuros de AEMET
      properties:
        nivel:
          $ref: '#/components/schemas/NivelRiesgo'
        comentarios:
          type: string
          example: "Predicción para mañana. Se evaluaron 3 variedades de cultivo."
        alertas:
          type: array
          items:
            $ref: '#/components/schemas/AlertaDTO'
        contexto:
          $ref: '#/components/schemas/ContextoCalculoDTO'
        tipo_prediccion:
          type: string
          enum: [hoy, manana, pasadomanana]
        evaluaciones_variedades:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/ResumenCultivoDTO'
        riesgos_heladas_blancas:
          type: array
          items:
            $ref: '#/components/schemas/RiesgoHeladaTipoDTO'
        riesgos_heladas_negras:
          type: array
          items:
            $ref: '#/components/schemas/RiesgoHeladaTipoDTO'
        precision:
          $ref: '#/components/schemas/TipoPrecision'
        datos_meteorologicos:
          $ref: '#/components/schemas/DatoAEMETDTO'
        evaluacion_localidades:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/ResumenEvaluacionLocalidadDTO'

    # Plagas
    CultivoDTO:
      type: object
      properties:
        nombre:
          type: string
          example: "Vid"
        grupo:
          type: string
          example: "Frutales"
...
    RiesgoPlagaCultivoDTO:
      type: object
      description: Riesgo de plagas calculado para un cultivo en la semana actual
      properties:
        cultivo:
          $ref: '#/components/schemas/CultivoDTO'
        plagas:
          type: array
          items:
            $ref: '#/components/schemas/PlagaDTO'
...
    RiesgoPlagaEstimadaDTO:
      type: object
      description: Resultado de predicción estimada de plagas para un cultivo en un rango de fechas
      properties:
        cultivo:
          type: string
          example: "vid"
        fecha_inicio:
          type: string
          format: date
          example: "2025-04-01"
        fecha_final:
          type: string
          format: date
          example: "2025-04-30"
        plagas_evaluadas:
          type: array
          items:
            $ref: '#/components/schemas/PlagaEvaluadaDTO'